requirements.txt
annotated-types==0.7.0
backoff==2.2.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
coinbase-advanced-py==1.8.2
cryptography==45.0.4
idna==3.10
numpy==2.3.0
pandas==2.3.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==3.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.4
six==1.10.0
sortedcontainers==2.4.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
websocket_client==0.40.0
websockets==13.1
